+++ php/tests/Application/Command/ClassInfoTest/ClassConstantDefaultValue.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace A;
+
+class TestClass
+{
+    const TEST_CONSTANT = [1, 'test'];
+}