+++ php/tests/Application/Command/ClassInfoTest/ClassMethodNoDocblock.php
@@ -4,8 +4,6 @@ namespace A;
 
 class TestClass
 {
-    const TEST_CONSTANT = 3;
-
     protected $testProperty;
 
     public function testMethod($parameter)